requirements.txt
# 딥러닝 프레임워크
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.0.0  # YOLOv5

# 이미지 처리
opencv-python>=4.6.0
Pillow>=9.0.0
rembg>=2.0.50  # 배경 제거
mediapipe>=0.10.0  # 얼굴/체형 분석
scikit-image>=0.21.0

# 웹 프레임워크
streamlit>=1.25.0
flask>=2.2.0

# 데이터 처리
numpy>=1.21.0
pandas>=1.4.0

# CLIP 모델
transformers>=4.20.0
ftfy>=6.1.0
regex>=2022.0.0

# 유틸리티
python-dotenv>=0.19.0
tqdm>=4.64.0
psutil>=5.9.0

# 개발 도구
pytest>=7.0.0
black>=22.0.0
